Add new VIF type VIF_TYPE_TAP
This change adds a new VIF type, VIF_TYPE_TAP, whose meaning is that
the host side of a vNIC to/from a VM is - at least initially - a
simple TAP interface that is not plugged into a bridge or vSwitch.  A
Neutron agent running on the compute host may subsequently set up
routing or bridging for that TAP interface.

VIF_TYPE_TAP supports a 'mac_address' key, in the VIF details dict,
which allows Neutron to specify the MAC address for the host end of
the TAP interface.

Change-Id: Ifd63faaa1ef4e085d9baa640c17da856e0960992
Implements: blueprint vif-type-tap